Add zip file for posting on RDS website
Choosing to add a dedicated zip folder file to package the lowest barrier to entry and not potentially overwhelm newer users but still provide content for most users (advanced users can probably navigate to the Github). We will see if I keep this for additional workshops in the future or if I will find a different approach.